Edits to remove enforcing of string formatting.

+++ lib/improver/tests/ensemble_copula_coupling/ensemble_copula_coupling/test_ResamplePercentiles.py
@@ -106,10 +106,10 @@ class Test__add_bounds_to_percentiles_and_forecast_values(IrisTest):
         bounds_pairing = (-40, 50)
         lower_array = np.full(
             forecast_at_percentiles[:, 0].shape, bounds_pairing[0],
-            dtype=np.int32)
+            dtype=np.float32)
         upper_array = np.full(
             forecast_at_percentiles[:, 0].shape, bounds_pairing[1],
-            dtype=np.int32)
+            dtype=np.float32)
         plugin = Plugin()
         result = plugin._add_bounds_to_percentiles_and_forecast_at_percentiles(
             percentiles, forecast_at_percentiles, bounds_pairing)
